Add styles to match recent blogs section
