+++ src/components/Home.jsx
@@ -3,6 +3,7 @@ import Welcome from './Welcome';
 import About from './About';
 import Stats from './Stats';
 import RecentBlogs from './RecentBlogs';
+import RecentGalleries from './RecentGalleries';
 
 const Home = () => (
   <div className="home">
@@ -10,6 +11,7 @@ const Home = () => (
     <About />
     <Stats />
     <RecentBlogs />
+    <RecentGalleries />
   </div>
 );
